folder restructure, favicon, title, signup page
