+++ Client/src/app/_components/auth/login/LoginForm.jsx
@@ -1,0 +1,17 @@
+'use client'
+import React from 'react'
+
+const LoginForm = () => {
+  return (
+    <div>
+      <form>
+        <input type="text" placeholder="Username" />
+        <input type="password" placeholder="Password" />
+        <input type="password" placeholder="Password" />
+        <button type="submit">Login</button>
+      </form>
+    </div>
+  )
+}
+
+export default LoginForm
