+++ Client/src/app/layout.js
@@ -4,8 +4,8 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Grocery App",
+  description: "Grocery Management tool with built-in Recipe integration and macro counter",
 };
 
 export default function RootLayout({ children }) {
