+++ Client/tailwind.config.js
@@ -11,6 +11,7 @@ module.exports = {
         "gradient-radial": "radial-gradient(var(--tw-gradient-stops))",
         "gradient-conic":
           "conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))",
+        'login': "url('/images/ingredients-wallpaper-signup.jpg')"  
       },
     },
   },
